fix: critical damage in single target (#204)
This adjustment resolves an issue where certain abilities, particularly those affecting an area (AoE), would sometimes fail to inflict any harm if the individual casting the ability and the intended recipient were the sole beings visible on the screen. This happened as a result of incomplete transmission of vital impact values when only a single subject was located within the zone.