fix: the source of the corruption (#238)
* Vocation Trial Fix

Remove Jo Staff properly.

* Fix: The Source of the Corruption

Discarded the functions onmove items from player.lua because it just doesn't work.
Created a movement check for the turtle in liberty bay. (Secret Library Quest)
Created an action script to use guilt on the cult bosses.
Corrected the corpse from Dark Soul and IDs in codes.
Made vulnerable the corruptor of souls with the prooper cooldown to return to stage 1 if the guilt item is not used on him.
Created the spawns of Dark Souls in world.
Corrected the boss room position in order to let spawn the dark souls in the boss room.
Now this boss is playable! Have fun!

* Source of Corruption Loot + deSpawn Mechanic

Just added the boss loot and fixed the coords. for erase the yalahari boss after the fight.

* Update player.lua

* Updated Monsters

Updated the monsters file with lastest modifications from other Pull Requests.

* update-monsters-lastest

Removed crustacea gigantica because it's already included in raids.xml
Crustacea gigantica does not spawn as regular spawn in real tibia.