Fixes: Warth of the Emperor Quest and Firefighter Achievement loop (#276)
* Fixes: Warth of the Emperor Quest and Firefighter Achievement loop

Fixed a loop with firefighter achievement, ensures to get it only one time.
Fixed bugs with wrath of emperor quest, now you can complete it.
Corrected the name of NPC in spawn and file.

* Update Prepared Bucket.

Updated file with lua format. Plus I noticed that you need to put out 500 flames to get the achievement, so I set a counter for it.